+++ .gitignore
@@ -21,7 +21,8 @@ DerivedData
 *.ipa
 *.xcuserstate
 **/.xcode.env.local
-
+.env
+env
 # Android/IntelliJ
 #
 build/
@@ -72,3 +73,6 @@ yarn-error.log
 !.yarn/releases
 !.yarn/sdks
 !.yarn/versions
+
+#temporary file
+/receivedmessages.json
