+++ package.json
@@ -22,19 +22,25 @@
     "react-native": "0.75.3",
     "react-native-agora-chat": "^1.3.0-alpha.0",
     "react-native-audio-recorder-player": "^3.6.11",
+    "react-native-create-thumbnail": "^2.0.0",
     "react-native-document-picker": "^9.3.1",
+    "react-native-file-viewer": "^2.1.5",
     "react-native-fontawesome": "^7.0.0",
     "react-native-fs": "^2.20.0",
     "react-native-gesture-handler": "^2.20.0",
     "react-native-image-crop-picker": "^0.41.3",
     "react-native-image-picker": "^7.1.2",
+    "react-native-openanything": "^0.0.6",
     "react-native-raw-bottom-sheet": "^3.0.0",
     "react-native-reanimated": "^3.15.4",
     "react-native-safe-area-context": "^4.11.0",
     "react-native-screens": "^3.34.0",
+    "react-native-sound": "^0.11.2",
     "react-native-sound-player": "^0.14.3",
     "react-native-splash-screen": "^3.3.0",
     "react-native-vector-icons": "^10.2.0",
+    "react-native-video": "^6.6.4",
+    "react-native-webview": "^13.12.3",
     "tailwind": "^4.0.0"
   },
   "devDependencies": {
