+++ src/components/LogoutButton.jsx
@@ -1,15 +1,15 @@
 import React, { useContext } from 'react';
 import { TouchableOpacity, Image } from 'react-native';
 import { useNavigation } from '@react-navigation/native';
-import { logout } from '../agora/authAgora';
+import { logout } from '../agora/Group/helpers';
 import AgoraContext from '../context/AgoraContext';
 
 const LogoutButton = () => {
-  const { chatClient, isInitialized } = useContext(AgoraContext);
+  const { chatClient } = useContext(AgoraContext);
   const navigation = useNavigation(); // Access the navigation object
 
   const handleLogout = async () => {
-    await logout(isInitialized, chatClient);
+    await logout(chatClient);
     navigation.reset({
       index: 0,
       routes: [{ name: 'LoginScreen' }],
