+++ src/screens/GroupDetails.jsx
@@ -1,34 +1,34 @@
 import React, { useContext, useState, useEffect, useRef } from 'react';
-import { View, Text, FlatList, TextInput, TouchableOpacity, ActivityIndicator, Image } from 'react-native';
+import { View, Text, FlatList, TextInput, TouchableOpacity, ActivityIndicator, Image, Dimensions, ImageBackground } from 'react-native';
 import { useRoute } from '@react-navigation/native';
-import { sendmsg, receivemessages, registerMessageListener } from '../agora/groupManager';
+import { sendmsg, receivemessages, registerMessageListener } from '../agora/Group/helpers';
 
 import AgoraContext from '../context/AgoraContext';
 import UploadFiles from '../components/UploadFiles';
 import { ChatMessageType } from 'react-native-agora-chat';
 import RecordVoice from '../components/RecordVoice';
 
-import AudioRecorderPlayer from 'react-native-audio-recorder-player';
-
+import VideoMessage from '../components/VideoMessage';
+import ImageMessage from '../components/ImageMessage';
+import PlayVoiceMessage from '../components/PlayVoiceMessage';
+import FileMessage from '../components/FileMessage';
 
 const GroupDetails = ({ navigation }) => {
-  const { chatClient, isInitialized } = useContext(AgoraContext);
+  const { chatClient } = useContext(AgoraContext);
   const { groupId, groupName } = useRoute().params;
   const [message, setMessage] = useState('');
   const [chats, setChats] = useState([]);
   const [username, setUsername] = useState('');
-  const [isPlaying, setIsPlaying] = useState(false);
-  const [currentlyPlayingId, setCurrentlyPlayingId] = useState(null);
-  const [playTime, setPlayTime] = useState('00:00');
-  const audioRecorderPlayer = useRef(new AudioRecorderPlayer()).current;
   let filePath;
   let displayName;
   let width;
   let height;
   let duration;
+  let fileSize;
+  let thumbnailLocalPath;
 
   const fetchMessages = async () => {
-    const updatedMessages = await receivemessages(isInitialized, chatClient, groupId);
+    const updatedMessages = await receivemessages(chatClient, groupId);
     console.log("CHATS: ", updatedMessages)
     setChats(updatedMessages);
   };
@@ -38,7 +38,6 @@ const GroupDetails = ({ navigation }) => {
     chatClient.getCurrentUsername().then(setUsername).catch(console.log);
 
     const messageListener = registerMessageListener(chatClient, groupId, setChats);
-    console.log("messageListener Chats: ", chats)
     return () => {
       chatClient.chatManager.removeMessageListener(messageListener);
     };
@@ -56,9 +55,17 @@ const GroupDetails = ({ navigation }) => {
     });
   };
 
+  const setFileData = async (pickedDoc) => {
+    filePath = pickedDoc.fileUri;
+    displayName = pickedDoc.fileName;
+    fileSize = pickedDoc.fileSize;
+    await handleSendMessage({ messageType: ChatMessageType.FILE });
+  }
+
   const setImageData = async (pickedImage) => {
     filePath = pickedImage.fileUri;
     displayName = pickedImage.fileName;
+    fileSize = pickedImage.fileSize;
     width = 150;
     height = 150;
     await handleSendMessage({ messageType: ChatMessageType.IMAGE });
@@ -71,72 +78,65 @@ const GroupDetails = ({ navigation }) => {
     await handleSendMessage({ messageType: ChatMessageType.VOICE });
   }
 
+  const setVideoData = async (video) => {
+    filePath = video.fileUri;
+    displayName = video.fileName;
+    duration = video.duration;
+    await handleSendMessage({ messageType: ChatMessageType.VIDEO });
+  }
+
   const handleSendMessage = async ({ messageType = null }) => {
     let msg;
     if (messageType === ChatMessageType.TXT) {
       msg = await sendmsg({
-        isInitialized, chatClient, groupId, messageType,
+        chatClient, groupId, messageType,
         message
       });
       setChats(prevChats => [...prevChats, { id: msg.id, message: msg.message, sender: msg.sender, type: 'txt' }]);
       setMessage('');
     } else if (messageType === ChatMessageType.IMAGE) {
       msg = await sendmsg({
-        isInitialized, chatClient, groupId, messageType,
+        chatClient, groupId, messageType,
         filePath,
-        displayName,
-        width,
-        height,
+        displayName
       });
-      setChats(prevChats => [...prevChats, { id: msg.id, image: msg.image, sender: msg.sender, type: 'img' }]);
+      setChats(prevChats => [...prevChats, { id: msg.id, image: msg.image, sender: msg.sender, fileSize: msg.fileSize, type: 'img' }]);
     } else if (messageType === ChatMessageType.VOICE) {
       msg = await sendmsg({
-        isInitialized, chatClient, groupId, messageType,
+        chatClient, groupId, messageType,
         filePath,
         displayName,
+        fileSize,
         duration,
       });
-      setChats(prevChats => [...prevChats, { id: msg.id, voice: msg.voice, sender: msg.sender, time: msg.duration, type: 'voice' }]);
-    }
-
-  };
-
-  const handlePlayPause = async (item, Id) => {
-    setCurrentlyPlayingId(Id);
-    if (!isPlaying) {
-      // Start playing
-      setIsPlaying(true); //Now it's playing 
-      const result = await audioRecorderPlayer.startPlayer(item.voice);
-      audioRecorderPlayer.addPlayBackListener((e) => {
-        setPlayTime(audioRecorderPlayer.mmss(Math.floor(e.currentPosition / 1000)));
-        if (Math.ceil(e.currentPosition / 1000) === item.duration) {
-          setIsPlaying(false);
-          setPlayTime('00:00');
-          // console.log("Player Stopped.")
-          // console.log("isPlaying",isPlaying)
-          audioRecorderPlayer.stopPlayer();
-        }
+      setChats(prevChats => [...prevChats, { id: msg.id, voice: msg.voice, sender: msg.sender, duration: msg.duration, fileSize: msg.fileSize, type: 'voice' }]);
+    } else if (messageType === ChatMessageType.FILE) {
+      msg = await sendmsg({
+        chatClient, groupId, messageType,
+        filePath,
+        displayName,
+        fileSize
       });
-    } else {
-      // Pause playing
-      await audioRecorderPlayer.pausePlayer();
-      setIsPlaying(false);
+      setChats(prevChats => [...prevChats, { id: msg.id, localUrl: msg.localUrl, remoteUrl: msg.remoteUrl , sender: msg.sender, displayName: msg.displayName, fileSize: msg.fileSize, type: 'file' }]);
+    } else if (messageType === ChatMessageType.VIDEO) {
+      msg = await sendmsg({
+        chatClient, groupId, messageType,
+        filePath,
+        displayName,
+        thumbnailLocalPath,
+        duration,
+      });
+      setChats(prevChats => [...prevChats, { id: msg.id, video: msg.video, sender: msg.sender, displayName: msg.displayName, duration: msg.duration, type: 'video' }]);
     }
-  };
 
-  const formatTime = (duration) => {
-    const minutes = Math.floor(duration / 60); // Get the minutes
-    const seconds = duration % 60; // Get the remaining seconds
-    return `${String(minutes).padStart(2, '0')}:${String(seconds).padStart(2, '0')}`; // Format MM:SS
   };
-
 
   const renderChatItem = ({ item }) => {
     return (
       <View
-        className={`p-2 border-b rounded-2xl max-w-xs mb-4 ${item.sender === username ?
-          item.type !== 'img' ? 'bg-blue-600 self-end mr-4' : 'self-end mr-4' :
-          item.type !== 'img' ? 'bg-green-500 self-start ml-4' : 'self-start ml-4'}`}
+        className={`p-2 border-b-0 rounded-2xl max-w-xs mb-4 ${item.sender === username ?
+          item.type === 'txt' || item.type === 'voice' ? 'bg-blue-600 self-end mr-4' : 'self-end mr-4' :
+          item.type === 'txt' || item.type === 'voice' ? 'bg-green-500 self-start ml-4' : 'self-start ml-4'}`}
       >
         <Text className="text-white font-bold text-lg ml-2">{item.sender}</Text>
 
@@ -144,38 +144,14 @@ const GroupDetails = ({ navigation }) => {
         {item.type === 'txt' ? (
           <Text className="text-white">{item.message}</Text>
         ) : item.type === 'img' && item.image ? (
-          <Image
-            source={{ uri: item.image }}
-            style={{ width: 150, height: 150 }}
-            className="rounded-lg mt-2 bg-blue-400"
-          />
+          <ImageMessage item={item}/>
         )
           : item.type === 'voice' && item.voice ? (
-            <View className="flex-row items-center mt-1">
-              <TouchableOpacity
-                onPress={() => handlePlayPause(item, item.id)}
-                className=" rounded-full w-[150px] h-[22px] justify-center items-start"
-              >
-                {isPlaying && currentlyPlayingId === item.id ? (
-                  <Image className="ml-3" source={require('../../assets/icons/pause.png')} style={{ width: 15, height: 15 }} />
-                ) : (
-                  <Image className="ml-3" source={require('../../assets/icons/play-button.png')} style={{ width: 15, height: 15 }} />
-                )}
-              </TouchableOpacity>
-              {isPlaying && currentlyPlayingId === item.id ? (
-                <Text className="mr-6 text-white">
-                  {playTime}
-                </Text>
-              ) : !isPlaying ? (
-                <Text className="mr-6 text-white">
-                  {formatTime(item.duration)}
-                </Text>
-              ): (
-                <Text className="mr-6 text-white">
-                  {formatTime(item.duration)}
-                </Text>
-              )}
-            </View>
+            <PlayVoiceMessage item={item} />
+          ) : item.type === 'file' ? (
+            <FileMessage item={item} />
+          ) : item.type === 'video' ? (
+            <VideoMessage item={item}/>
           ) : null}
       </View>
     );
@@ -213,19 +189,18 @@ const GroupDetails = ({ navigation }) => {
           onChangeText={text => setMessage(text)}
           value={message}
         />
-        {/* Upload Files Component */}
-        <UploadFiles setImageData={setImageData} />
-        {/* Record Voice Message Component */}
-        <RecordVoice setVoiceMessageData={setVoiceMessageData} />
-
         <TouchableOpacity
           className="ml-2 bg-blue-500 p-2.5 rounded-full"
           onPress={() => handleSendMessage({ messageType: ChatMessageType.TXT })}
         >
           <Text className="text-black">Send</Text>
         </TouchableOpacity>
-      </View>
 
+        {/* Upload Files Component */}
+        <UploadFiles setImageData={setImageData} setFileData={setFileData} setVoiceMessageData={setVoiceMessageData} setVideoData={setVideoData} />
+        {/* Record Voice Message Component */}
+        <RecordVoice setVoiceMessageData={setVoiceMessageData} />
+      </View>
     </View>
   );
 };
